requirements.txt
Django==2.1.4
django-filter==2.1.0
djangorestframework==3.9.1
gunicorn==19.9.0
mysqlclient==1.3.14
pytz==2018.7
PyYAML==3.13
